Add a vcl-reload script that runs permanently to watch for changes.
This script uses inotifywait will block until a change is made, then the loop body will be executed.
Downside is that if a change is made during the body execution it will not be detected.

Another way to code this is doing a *inotifywait -m | while read [...]*
but then you need a way to avoid running multiple compile because all files were changed at once.
So the simplest and more robust way was choosen.